% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torscrape.R
\name{scrape}
\alias{scrape}
\title{scrape webpages with Tor proxy}
\usage{
scrape(urls, async = FALSE, max_retry = 3, random_ua = TRUE, callback = NULL)
}
\arguments{
\item{urls, }{character vector, a character vector of urls to be scraped}

\item{async, }{bool (default FALSE), to use async request or not}

\item{max_retry, }{int (default 5), how many times to allow for repeating}

\item{random_ua, }{bool (default TRUE), whether to use Randomuseragent::random_useragent() or not}

\item{callback, }{function (default return entire request contenty), take request content as input and return a single character as output}
}
\description{
scrape webpages with Tor proxy
}
